checks for handling edge_bits to avoid allocations
The claim is that the extra memory in edge_bits is more than enough to
be (re)used for bitsets that are part of the clique search. Seems solid,
but there's a condition in the graph ctor to check if the memory will be
enough and extend if needed (unlikely it will ever be called).

check_memory() calls are moved to the end of a potential search for the
DFS methods, and once at the start for the heuristic. Cleaning out the
memory for reuse is amortized across all searches. None of the examples
require more than the available memory to complete the search.

This means that there will be no heap allocations at the most frequent
location earlier: the clique search. All allocations are part of
initializing the objects.