fix candidate counting error in dfs methods
an error was introduced while updating from v1.2.0 to v1.3.0 where the
clique_potential (and correct set of candidates) was not updated
appropriately after a previous search was avoided. similar error was
fixed in enumerate. some cosmetic updates to clique and dfs_recurse,
need to check recurse for similar counting errors.