a simple heuristic to prune unnecessary branches
since the radix sort offers a small speed boost when loading the graph,
plus the no memory allocations during searches means we can spend a
little time computing a simple heuristic: number of neighbors of a
vertex that have same or higher degree. This slows down the loading a
little (need to go through all edges), but gives an upper_bound on the
clique size and speeds up the heuristic/dfs by giving a stricter pruning.