chnaged enumeration test for correspondences
since the test graph for finding correspondences is generated from a
random set of points, occasionally the enumerations end in errors (ie
there may be more cliques of size 9, even if the maximum is 10; or the
maximum may even not be 10). The test now depends on the size of the
maximum rather than a hardcoded 10.