lots of small changes
remove the block_empty search in dfs_recurse, we can add CLZ/CTZ if
needed later on.

remove the dlen data member from gbits struct, it is rarely used and we
can always compute it if needed.

avoid writing the mcs value during the heuristic, it may be that the
actual clique is found at CUR_MAX_CLIQUE_LOC.