avoid unnecessary binary search
the OUTSIDE result for binary search was useful in reducing the number
of times the routine had to be called, but it slowed overall due to an
external check every time.

instead of that, binary_find now returns the start, end, or correct
index to the referenced loc variable which means that after one
OUTSIDE-equivalent search, the remaining binary_finds have to search an
array of length 1.