fix binary search change
the previous change does not improve performance because we are shifting
the start of the array and its length -- so when ans is overwritten with
N-1, in the next call it is overwritten with 1, which means the binary
search goes through a larger space then necessary.

the fix is to have an additional variable called start, which allows to
set the start of the array for binary_find, and start accumulates the
value of ans. note that binary_find gives valid value of ans in all
cases now, and the end check is shifted first.