added node-weighted clique search methods
the heuristic/DFS/enumerate work similarly to the unweighted case,
although there may be some additional memory usage due to handling the
weights. memory access in general may be whack because every time I add
or remove a vertex from the clique I have to access its weight.

weights are stored as doubles, the nwgraph ctor needs a vector of
weights. currently the sentinel weight is zero, find a way to wrap it
better.

perhaps adding and subtracting doubles isn't as nice as integers, it may
be worth it (accuracy-wise) to use a loop for clique_potential instead.

added another file to mem_test to check the usage for the node-weighted
case -- no memory issues detected by valgrind as part of this commit.
the heuristic/DFS/enumerate methods provide the correct answer for the
simple example.

the python wrapper has not been implemented yet. the "speed" of this
algorithm wrt the unweighted method has not been tested.