let the STL handle graph memory within pygraph
earlier I used a raw pointer for the graph object within the pygraph
object and I handled freeing that graph memory manually. I resorted to
this because the internal implementation of graph was roundabout (gbits
had rule of 5 with copy ctors and move ctors, but vertex had none of
them) and the compiler complained when I tried to use shared_ptr or
unique_ptr. But now:

* I don't (want to) remember how the manual memory handling works
  anymore: initialization is a one-time thing, and copies (within C++
  and/or outside pybind11's refcount) are pretty rare.

* I tested that shared_ptr does the right thing when it comes to moves
  and copies with a simple example.

* the internals of graph are straightforward now: only STL containers
  are used, there are no weird ctors/operator= magics in the background,
  and the raw pointers within gbits function like offsets: they don't
  own memory, are never alloc'ed or freed.

Thus the raw pointer and ctor shenanigans within pygraph are now gone,
replaced by a shared_ptr.

* the clique search is unchanged due to any of this, perhaps using a
  shared_ptr might slowdown calling the member functions, but that's a
  one-time thing as well. So I don't expect this to affect the speed.

* There's a hit to compilation time because no PIMPL, but that's more
  because python's setup.py can't compile object files in parallel. Oh
  well.