added python wrapper for nwgraph
creates a class called NWGraph in Python. mostly similar to Graph, but
no time limits, and you have to pass vertex weights for construction.

there is some code duplication, but its ok for now, I'm not sure if we
can extract out the "common" functionality.

factory methods are ok, file IO is not implemented. time limits are not
provided, reset_search looks unnecessary as well.